+++ src/convert_roms/test_original_to_zip.rs
@@ -67,6 +67,7 @@ async fn test() {
         games_by_id,
         romfiles_by_id,
         false,
+        1,
         false,
     )
     .await