+++ src/sort_roms/test_sort.rs
@@ -66,8 +66,8 @@ async fn test() {
         &one_regions,
         &[],
         &[],
-        "NONE",
-        "NONE",
+        "none",
+        "none",
     )
     .await
     .unwrap();