+++ src/sort_roms/test_sort_1g1r_catch_all.rs
@@ -68,8 +68,8 @@ async fn test() {
         &one_regions,
         &[],
         &[],
-        "NONE",
-        "NONE",
+        "none",
+        "none",
     )
     .await
     .unwrap();