+++ src/sort_roms/test_sort_1g1r_discard_asia_and_beta.rs
@@ -68,8 +68,8 @@ async fn test() {
         &one_regions,
         &["Beta"],
         &[],
-        "NONE",
-        "NONE",
+        "none",
+        "none",
     )
     .await
     .unwrap();