+++ src/sort_roms/test_sort_1g1r_subfolders_alpha.rs
@@ -67,8 +67,8 @@ async fn test() {
         &one_regions,
         &[],
         &[],
-        "NONE",
-        "ALPHA",
+        "none",
+        "alpha",
     )
     .await
     .unwrap();